Further refactoring of adjust_surface_pressure
The old version used loops and was extremely slow. The routine was refactored using various features of xarray to eliminate all loops. for a ne30pg2 grid case this brought the execution time down dramatically from 4 minutes to 1 sec. Further refactoring should be undertaken to break up this routine to allow for more torough unit testing.

	modified:   create_initial_condition.py
	modified:   hiccup_data_class.py
	modified:   hiccup_state_adjustment.py
	modified:   plot.sanity_check.py
	modified:   test_state_adjustment.py