Fixed handling of datasets with "with" statements
previous appraoch was to load and dataset in the conventional way, then write back to the netcdf file and "close" the dataset object, but this causes a strange problem when you try to do it twice in the same script. Xarray docs mention that it is generally safer to use a "with" clause, and sure enough it fixed the issue here!

	modified:   create_initial_condition.py