Add logic for ncremap lrg2sml flag
Added regex lines to determine the "ne" value (number of elements on
cube edge) and use this number to determine if the lrg2sml flag should
be set. This is sort of a hack because it depends on the relative grid
size between the input and output. More work will be needed to
dynamically determine which grid is bigger. We'd probably want to read
each grid file and determine an average spacing, or something like that.

	modified:   hiccup_data_class.py