Bug fix for remap_vertical()
The command to "mv" the temporary file at the end of the vertical remap
routine was causing a strange error. It appeared to be that the file was
not completely written before accessing it again in the next step
(second state adjustment). However, a brute force "sleep" command did
not resolve the issue. After further experimentation opening and closing
a xarray dataset solved the issue, but I do not understand why.

	modified:   hiccup_data_class.py