Add larger file header and new multifile methods
Both of these are aimed at creating better support for large
filerequired by very fine grids like ne1024. It appears the multi-file
method is crucial for this because most machine simply do not have
enough memory to operate on a single 100-200GB file with all variables,
so the multi-file methods are aimed at an alternate workflow where the
data is brokedn out in a single file for each variable. More methods are
needed to complete this alternate workflow.

	modified:   hiccup_data_class.py