Clean up scripts
The multi-file workflow was working so well that I decided to collapse
all the "create" scripts into a single script that will work for all
resolutions. The original single file workflow is probably better suited
for low-res cases, which will be the majority of use cases, but trying
to support two very different workflow seemed very hard to do in this
case. I'm open to revisiting this design decision later on. In the
meantime I'm going to leave all the hiccup_data_class routines in place,
even if they aren't used by the multifile workflow.

	deleted:    create_initial_condition.nersc.py
	deleted:    create_initial_condition.olcf.py
	renamed:    create_initial_condition_multifile.nersc.py -> create_initial_condition.py
	deleted:    create_initial_condition_multifile.olcf.py