Update surface pressure adjustment
The ne1024 case revealed a problem with the surface pressure adjustment
that produced inf values in the output. This was fixed by limiting small
or negative surface topo values in the limiting conditions for alpha.
These conditional blocks are really only menat for certain conditions on
elevated terrain, so this shouldn't have any ill effects and only seems
to apply to these very high resolution datasets.

	modified:   hiccup_state_adjustment.py