Fix broadcasting of PHIS in state adjustments
Fix broadcasting of PHIS in state adjustment routines. An issue arises
when PHIS has shape (ncol), but PS has shape (ntime, ncol). In this
case, the broadcasting of computed arrays depends on the order in which
these arrays appear in computations. So if PS appears first, then
computed arrays will have shape (ntime, ncol), but if PHIS appears
first, computed arrays will have shape (ncol, ntime). The if the values
attributed is used to explicitly access the underlying data, we can get
mismatched shapes that end up with numpy broadcasting to shape (ncol,
ncol). To prevent this, we can remove the direct access to values so
that xarray will always do smart broadcasting, but to ensure that
computed arrays end up with our desired shape of (ntime, ncol) as
opposed to (ncol, ntime) we also explicitly broadcast del_phis to the
shape of PS.