Make hiccup an installable package
Make hiccup an installable package. This allows installing hiccup via
setup.py (and should also allow installing via pip), and then being able
to import the hiccup package from any directory. The directory structure
needed to be changed slightly to accommodate this. The hiccup source is
moved to the hiccup directory, and unit tests are moved to the tests
directory. The scripts are modified accordingly to correctly import the
hiccup modules. For example, `import hiccup_data_class as hdc` needed to
be renamed to `from hiccup import hiccup_data_class as hdc` since the
location of the source has moved. This is also consistent with `hiccup`
being the package name encompassing all of these tools, so running the
codes from the root of this project tree should not care if hiccup has
been installed as a python package or not.